Removed unused code and fixed caching.

+++ QuestPackages/Services/CachingService.cs
@@ -12,7 +12,7 @@ namespace QuestPackages.Services
         private readonly PackageAPIService _packageAPIService;
         private readonly PackageDBService _packageDBService;
         private List<string> _packagesIdsToDelete { get; set; }
-        private List<string> _dbPackageIds { get; set; }
+        private List<string> _packageIds { get; set; }
         public CachingService(PackageAPIService packageAPIService, PackageDBService packageDBService) 
         {
             _packageAPIService = packageAPIService;
@@ -22,7 +22,7 @@ namespace QuestPackages.Services
         public async Task UpdateCache() 
         {
             await ComparePackages();
-            await UpdatePackages(_dbPackageIds);
+            await UpdatePackages(_packageIds);
             DeletePackages();
         }
 
@@ -37,7 +37,6 @@ namespace QuestPackages.Services
 
                 Package package = await _packageAPIService.GetPackage(packageId, packageVersions[0]);
 
-
                 dbPackage = _packageDBService.PackageToDBPackage(package);
                 dbPackage.Versions = packageVersions.ToArray();
                 if (_packageDBService.HasPackage(packageId)) _packageDBService.UpdatePackage(packageId, dbPackage);
@@ -47,7 +46,7 @@ namespace QuestPackages.Services
 
         public void DeletePackages() 
         {
-            if (_packagesIdsToDelete.Count == 0) return;
+            if (_packagesIdsToDelete == null || _packagesIdsToDelete.Count == 0) return;
             foreach(var packageId in _packagesIdsToDelete) 
             {
                 _packageDBService.DeletePackage(packageId);
@@ -56,10 +55,10 @@ namespace QuestPackages.Services
 
         public async Task ComparePackages() 
         {
-            _dbPackageIds = await _packageAPIService.GetPackageIds();
+            _packageIds = await _packageAPIService.GetPackageIds();
             List<string> dbPackageIds = _packageDBService.GetPackageIds();
 
-            _packagesIdsToDelete = _dbPackageIds.Except(dbPackageIds).ToList();
+            _packagesIdsToDelete = dbPackageIds.Except(_packageIds).ToList();
         }
     }
 }
